road-classification/requirements.txt
LatLon==1.0.2
numpy==1.14.2
pyproj==1.9.5.1
python-dateutil==2.7.2
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
